fixed deleting dirs

+++ netcdu/Program.cs
@@ -41,8 +41,8 @@ F4 - Closes the application
             var fs = new FileScanner(path, new DefaultGetFileSizeStrategy());
 
             Application.Init();
-            Colors.Base.Normal      = Application.Driver.MakeAttribute(Color.White, Color.Black);
-            Colors.Base.HotNormal   = Application.Driver.MakeAttribute(Color.White, Color.Black);
+            Colors.Base.Normal = Application.Driver.MakeAttribute(Color.White, Color.Black);
+            Colors.Base.HotNormal = Application.Driver.MakeAttribute(Color.White, Color.Black);
             var top = Application.Top;
             var window = new Window(new Rect(0, 1, top.Frame.Width, top.Frame.Height - 2), "netcdu");
             var tree = new TreeView
@@ -55,17 +55,17 @@ F4 - Closes the application
                 AllowsMultipleSelection = false
             };
             window.Add(tree);
-            var statusBar = new StatusBar(new [] {
+            var statusBar = new StatusBar(new[] {
             new StatusItem(Key.F1, "~F1~ Help", Help ),
             new StatusItem(Key.F2, "~F2~ Go up",async() => await GoUp(window,fs,top,tree) ),
             new StatusItem(Key.F3, "~F3~ Delete", () => Delete(tree) ),
             new StatusItem(Key.F4, "~F4~ Exit", Application.RequestStop ),
             });
             var menu = new MenuBar();
-            
+
             top.Add(statusBar, window, menu);
 
-            top.Ready += async()=>await ScanFiles(window,fs,top, tree);
+            top.Ready += async () => await ScanFiles(window, fs, top, tree);
             Application.Run(top);
 
             Console.Clear();
@@ -157,10 +157,10 @@ F4 - Closes the application
 
         static async Task GoUp(Window window, FileScanner fs, Toplevel top, TreeView tree)
         {
-            if(!fs.GoUp())
+            if (!fs.GoUp())
                 return;
             tree.Root = null;
-            await ScanFiles(window, fs, top,tree);
+            await ScanFiles(window, fs, top, tree);
         }
 
         static void Delete(TreeView tv)
@@ -169,7 +169,11 @@ F4 - Closes the application
             if (index == 1)
             {
                 ITreeViewItem nodeToDelete = tv.SelectedItem;
-                File.Delete(nodeToDelete.ToString());
+                if (File.Exists(nodeToDelete.ToString()))
+                    File.Delete(nodeToDelete.ToString());
+                else
+                if (Directory.Exists(nodeToDelete.ToString()))
+                    Directory.Delete(nodeToDelete.ToString(), true);
                 ((INetcduNode)nodeToDelete).Delete();
 
                 tv.SelectedItem = nodeToDelete.Parent;
